Splitted aadToken to different variable provider, support for new params
