+++ src/models/variableType.ts
@@ -1,6 +1,7 @@
 "use strict";
 
 export enum VariableType {
+    AadToken,
     Environment,
     File,
     Request,
